requirements.txt
langchain-community
langchain-openai
tiktoken
langchainhub
chromadb
langchain
langgraph
tavily-python
gpt4all
python-dotenv
beautifulsoup4
unstructured[all-docs]
pypdf